[misc] fix clang warnings

+++ libs/exo/include/exo/collections/dynamic_array.h
@@ -58,7 +58,7 @@ struct DynamicArray
 	constexpr T       *begin() noexcept { return &array[0]; }
 	constexpr const T *end() const noexcept { return begin() + array_size; }
 	constexpr T       *end() noexcept { return begin() + array_size; }
-	constexpr usize    size() const noexcept { return end() - begin(); }
+	constexpr usize    size() const noexcept { return usize(end() - begin()); }
 	constexpr usize    capacity() const noexcept { return CAPACITY; }
 	constexpr bool     empty() const noexcept { return array_size == 0; }
 	constexpr const T *data() const noexcept { return &array[0]; }
