+++ libs/exo/include/exo/maths/u128.h
@@ -1,5 +1,5 @@
 #pragma once
-#include <exo/maths/numerics.h>
+#include "exo/maths/numerics.h"
 
 #include <emmintrin.h>
 #include <smmintrin.h>
@@ -11,11 +11,11 @@ using u128 = __m128i;
 #if __x86_64__ || _M_AMD64
 inline void u128_to_u64(u128 value, u64 *val0, u64 *val1)
 {
-	*val0 = _mm_extract_epi64(value, 0);
-	*val1 = _mm_extract_epi64(value, 1);
+	*val0 = u64(_mm_extract_epi64(value, 0));
+	*val1 = u64(_mm_extract_epi64(value, 1));
 }
 
-inline u128 u128_from_u64(u64 val1, u64 val0) { return _mm_set_epi64x(val1, val0); }
+inline u128 u128_from_u64(u64 val1, u64 val0) { return _mm_set_epi64x(i64(val1), i64(val0)); }
 #else
 #error Cannot determine architecture to use!
 #endif
